+++ apps/main-app/ts/features/itwallet/common/saga/__tests__/linking.test.ts
@@ -1,9 +1,9 @@
 import { testSaga } from "redux-saga-test-plan";
 
-import { handleItwStoredDeepLink } from "..";
 import NavigationService from "../../../../../navigation/NavigationService";
-import { ITW_ROUTES } from "../../routes";
-import { parseItwDeepLink } from "../../utils";
+import { ITW_ROUTES } from "../../../navigation/routes";
+import { parseItwDeepLink } from "../../utils/linking";
+import { handleItwStoredDeepLink } from "../linking";
 
 describe("handleItwStoredDeepLink", () => {
   it("navigates to a regular ITW route", () => {
