+++ apps/main-app/ts/features/itwallet/common/utils/__tests__/linking.test.ts
@@ -1,4 +1,4 @@
-import { parseItwDeepLink } from "..";
+import { parseItwDeepLink } from "../linking";
 
 describe("parseItwDeepLink", () => {
   test.each([
