+++ apps/main-app/ts/features/itwallet/common/utils/linking.ts
@@ -5,8 +5,8 @@ import {
   IO_UNIVERSAL_LINK_PREFIX
 } from "../../../../utils/navigation";
 import { extractPathFromURL } from "../../../../utils/url";
+import { itwLinkingConfig } from "../../navigation/linking";
 import { parseCredentialOfferLink } from "../../offer/utils";
-import { itwLinkingConfig } from "../linking";
 
 const itwLinkPrefixes = [IO_INTERNAL_LINK_PREFIX, IO_UNIVERSAL_LINK_PREFIX];
